fix: refactored whitebottompart to accomodate 2 versions
